requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
mysql-connector-python==8.2.0
python-dotenv==1.0.0
bcrypt==4.1.2
PyMySQL==1.1.0
cryptography==41.0.7
